Commited By PR Chauhan

+++ Electra_WebApi/Controllers/CityMVCController.cs
@@ -12,6 +12,7 @@ namespace Electra_WebApi.Controllers
     public class CityMVCController : Controller
     {
         HttpClient client = new HttpClient();
+        private CraModel db = new CraModel();
         // GET: CityMVC
         public ActionResult Index()
         {
@@ -51,6 +52,14 @@ namespace Electra_WebApi.Controllers
         // GET: CityMVC/Create
         public ActionResult Create()
         {
+           
+            ViewBag.CL = GetStateList();
+
+            return View();
+        }
+
+        public List<State> GetStateList()
+        {
             List<State> list = new List<State>();
             client.BaseAddress = new Uri("https://localhost:44305/api/StateApi");
             var response = client.GetAsync("StateApi");
@@ -63,11 +72,8 @@ namespace Electra_WebApi.Controllers
                 display.Wait();
                 list = display.Result;
             }
-            ViewBag.CL = list;
-
-            return View();
+            return list;
         }
-
         // POST: CityMVC/Create
         [HttpPost]
         public ActionResult Create(City collection)
@@ -78,6 +84,14 @@ namespace Electra_WebApi.Controllers
                 collection.DoM = DateTime.Now;
                 collection.E_UserID = "admin";
                 collection.M_UserID = "admin";
+                if (!Validate(collection.City_Name))
+                {
+                    ViewBag.CL = GetStateList();
+                    ModelState.AddModelError(nameof(City.City_Name), "Duplicate City is not allowed..!!");
+                    return View(collection);
+
+                }
+               
                 client.BaseAddress = new Uri("https://localhost:44305/api/CityApi");
                 client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
                 var putdata = client.PostAsJsonAsync("CityApi", collection);
@@ -197,8 +211,20 @@ namespace Electra_WebApi.Controllers
             try
             {
                 // TODO: Add delete logic here
+                var st = (from l in db.Consignees
+                          where l.City_ID == id
+                          select new
+                          {
+                              sName = l.City_ID
+                          }).ToList();
+                if (st.Count > 0)
+                {
+                    ModelState.AddModelError(nameof(City.City_Name), "City is used in Consignees master, So city can't be delete..!!");
+                    return View(collection);
+                }
 
-                City lstCity = null;
+
+
                 client.BaseAddress = new Uri("https://localhost:44305/api/CityApi/");
                 var response = client.DeleteAsync("CityApi?id=" + id.ToString());
                 response.Wait();
@@ -214,6 +240,36 @@ namespace Electra_WebApi.Controllers
             {
                 return View();
             }
+        }
+
+        public bool Validate(string Parameter)
+        {
+            bool lRetVal = true;
+            //HttpClient _client = new HttpClient();
+            //List<City> list = new List<City>();
+            //_client.BaseAddress = new Uri("https://localhost:44305/api/CityApi");
+            //var response = _client.GetAsync("CityApi");
+            //response.Wait();
+
+            //var test = response.Result;
+            //if (test.IsSuccessStatusCode)
+            //{
+            //    var display = test.Content.ReadAsAsync<List<City>>();
+            //    display.Wait();
+            //    list = display.Result;
+            //}
+
+            var st = (from l in db.Cities
+                      where l.City_Name == Parameter
+                      select new
+                      {
+                          sName = l.City_Name
+                      }).ToList();
+            if (st.Count > 0)
+            {
+                lRetVal = false;
+            }
+            return lRetVal;
         }
     }
 }
